requirements.txt
boto3>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0
duckdb>=0.10.0
streamlit>=1.35.0
plotly>=5.22.0
